Move AdventurePlatformImpl to impl-common
